requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cached-property==2.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
decorator==5.2.1
fastapi==0.121.2
flatbuffers==25.9.23
frozenlist==1.8.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.26.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-dotenv==1.2.1
python-multipart==0.0.20
python-stdnum==2.1
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
SQLAlchemy-Ext==0.2
sqlalchemy-orm==1.2.10
starlette==0.49.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tornado==6.5.2
trio==0.32.0
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0
zope.interface==8.1.1
